Big update.
Added new class - its name is "not_a_hole" and it is very useful cauze when I'm trying to detect holes with OpenCV, some of images are just parts of that holes and it was annoying. Now that images just go to the "not_a_hole" folder and not processed. So cauz of this I also update dataset and it'll be next commit